Correcion de errores (#89)
* Solucionado fallo de añadir accesorio a carrito

El problema era en la vista, que debido a numerosos cambios, nos equivocamos al declarar correctamente el valor del id. Ahora ya se realiza correctamente, la operación.

* Correcciones de eliminar (En proceso)

Ya se ha solucionado al eliminación de catergoría. De producto esta en proceso.

* Todas las correciones realizadas.

Correciones de los diferentes fallos corregidas, ademas de añadido todos los mensajes de cuando realiza o no una accion correctamente.

+++ CarUPO/app/Http/Controllers/CategoriasController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Categoria;
+use App\Models\Producto_categoria;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
@@ -37,17 +38,28 @@ class CategoriasController extends Controller
         $categoria = new Categoria();
         $categoria->nombre = $request->nombre;
         $categoria->save();
-        return app()->make(CategoriasController::class)->callAction('mostrarCategorias', []);
+
+        $categorias = Categoria::all();
+        $success = "Categoria creada correctamente.";
+
+        return view('categorias', @compact('categorias', 'success'));
     }
 
 
     public function removeToCategorias(Request $request)
     {
         $categoria = Categoria::findOrFail($request->id);
+
+        if ($categoria->productos_categorias->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar la categoría porque está asociada a productos.');
+        }
         $categoria->delete();
-        return app()->make(CategoriasController::class)->callAction('mostrarCategorias', []);
+        $categorias = Categoria::all();
+        $success = "Categoria eliminada correctamente.";
+
+        return view('categorias', @compact('categorias', 'success'));
     }
-    
+
     public function editarCategoria(Request $request)
     {
 
@@ -71,6 +83,11 @@ class CategoriasController extends Controller
         $categoria = Categoria::findOrFail($request->id);
         $categoria->nombre = $request->nombre;
         $categoria->save();
-        return app()->make(CategoriasController::class)->callAction('mostrarCategorias', []);
+
+
+        $categorias = Categoria::all();
+        $success = "Categoria actualizada correctamente.";
+
+        return view('categorias', @compact('categorias', 'success'));
     }
 }
