+++ CarUPO/app/Http/Controllers/CochesController.php
@@ -89,7 +89,9 @@ class CochesController extends Controller
         $coche->nPuertas = $request->nPuertas;
         $coche->fk_producto_id = $producto->id;
         $coche->save();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $productos = Producto::paginate(8);
+        $success = "Coche creado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
 
     public function verBorrarCoche(Request $request)
@@ -194,12 +196,29 @@ class CochesController extends Controller
         $coche->nPuertas = $request->nPuertas;
         $coche->fk_producto_id = $producto->id;
         $coche->save();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $productos = Producto::paginate(8);
+        $success = "Coche editado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
     public function eliminarCoche(Request $request)
     {
+
         $coche = Coche::findOrFail($request->id);
+        $producto = $coche->producto;
+        if ($producto->lineas_de_compra->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociado a lineas de compra.');
+        }
+        if ($producto->lineas_de_carrito->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociado a lineas de carrito.');
+        }
+        if ($producto->favoritos_productos->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociada a favoritos.');
+        }
         $coche->delete();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $producto->delete();
+
+        $productos = Producto::paginate(8);
+        $success = "Coche eliminado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
 }
